Tests/LibWeb/Text/expected/navigation/location-reload-srcdoc.txt

iframe is loaded
iframe is going to reload
iframe is loaded
